Move getCandidates to the shared directory
